refactor: moved Order Book types and interface
